% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ci_glm.R
\name{add_ci.glm}
\alias{add_ci.glm}
\title{Confidence Intervals for Expected Response of Generalized Linear Models}
\usage{
\method{add_ci}{glm}(tb, fit, alpha = 0.05, names = NULL, response = TRUE,
  type = "parametric", ...)
}
\arguments{
\item{tb}{A tibble or Data Frame.}

\item{fit}{An object of class \code{glm}.}

\item{alpha}{A real number between 0 and 1. Controls the confidence
level of the interval estimates.}

\item{names}{NULL or character vector of length two. If
\code{NULL}, confidence bounds will automatically be named by
\code{add_ci}, otherwise, the lower confidence bound will be
named \code{names[1]} and the upper confidence bound will be
named \code{names[2]}.}

\item{response}{A logical. If \code{TRUE}, the confidence intervals
will be determined for the expected response, if \code{FALSE},
confidence intervals will be made on the scale of the linear
predictor.}

\item{type}{A string, either \code{"parametric"} or
\code{"bootstrap"}.  but at the moment, the bootstrap method is
not implemented.}

\item{...}{Additional arguments.}
}
\value{
A tibble, \code{tb}, with predicted values, upper and lower
    confidence bounds attached.
}
\description{
This function is one of the methods for \code{add_ci}, and is
called automatically when \code{add_ci} is used on a \code{fit} of
class \code{glm}. Confidence Intervals are determined by making an
interval on the scale of the linear predictor, then applying the
inverse link function from the model fit.
}
