% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_probs_lm.R
\name{add_probs.lm}
\alias{add_probs.lm}
\title{Response Probabilities for Linear Models}
\usage{
\method{add_probs}{lm}(tb, fit, q, name = NULL, comparison = "<",
  log_response = FALSE, ...)
}
\arguments{
\item{tb}{A tibble or Data Frame on which to append probabilities}

\item{fit}{An object of class lm. Predictions are made with this
object.}

\item{q}{A double. A quantile of the response distribution.}

\item{name}{NULL or character vector of length one. If \code{NULL},
probabilities will automatically be named by
\code{add_probs()}, otherwise, the probabilities will be named
\code{name} in the returned tibble}

\item{comparison}{\code{"<", or ">"}. If \code{comparison = "<"},
then Pr(Y|x < q) is calculated for each observation in
\code{tb}. Otherwise, Pr(Y|x > q) is calculated.}

\item{log_response}{A logical. Default is \code{FALSE}, set to
\code{TRUE} if the model is log-linear: \eqn{\log{Y} = X \beta
+ \epsilon}.}

\item{...}{Additional arguments.}
}
\value{
A tibble, \code{tb}, with predicted values and
    probabilities attached.
}
\description{
This is the method \code{add_probs} uses if the model fit is
linear.  Probabilities are calculated parametrically, using a
pivotal quantity.
}
\examples{
fit <- lm(dist ~ speed, data = cars)
add_probs(cars, fit, q = 20)
add_probs(cars, fit, q = 30, comparison = ">")

}
\seealso{
\code{{\link{add_ci.lm}}} for confidence intervals for
    \code{lm} objects. \code{\link{add_pi.lm}} for prediction
    intervals of \code{lm} objects, and
    \code{\link{add_quantile.lm}} for response quantiles of
    \code{lm} objects.
}
