% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_probs_lmer.R
\name{add_probs.lmerMod}
\alias{add_probs.lmerMod}
\title{Response Probabilities for Linear Mixed Models}
\usage{
\method{add_probs}{lmerMod}(tb, fit, q, name = NULL, comparison = "<",
  type = "parametric", includeRanef = TRUE, nSims = 200,
  log_response = FALSE, yhatName = "pred", ...)
}
\arguments{
\item{tb}{A tibble or Data Frame.}

\item{fit}{An object of class \code{lmerMod}.}

\item{q}{A double. A quantile of the response variable}

\item{name}{NULL or character vector of length two. If \code{NULL},
probabilities will automatically be named by \code{add_pi},
otherwise, the probabilities will be named \code{name} in the
returned tibble.}

\item{comparison}{A character vector of length one. Must be either
\code{"<"} or \code{">"}. If \code{comparison = "<"}, then
\eqn{Pr(Y|x < q)} is calculated for each x in the new data,
\code{tb}. Otherwise, \eqn{Pr(Y|x > q)} is calculated.}

\item{type}{A string, either \code{"parametric"} , \code{"sim"}, or
\code{"boot"}.}

\item{includeRanef}{A logical. Set whether the predictions and
intervals should be made conditional on the random effects. If
\code{FALSE}, random effects will not be included.}

\item{nSims}{A positive integer. If \code{type = "sim"}
\code{nSims} will determine the number of simulated draws to
make.}

\item{log_response}{A logical. Set to \code{TRUE} if your model is
a log-linear mixed model.}

\item{yhatName}{A string. Determines the name of the vector of
predictions.}

\item{...}{Additional arguments.}
}
\value{
A tibble, \code{tb}, with predictions and probabilities
    attached.
}
\description{
This function is one of the methods for \code{add_probs}, and is
called automatically when \code{add_probs} is used on a \code{fit}
of class \code{lmerMod}.
}
\details{
It is recommended that one perform a parametric bootstrap to
determine these probabilities. To do so, use the option \code{type = "boot"}
}
\examples{
dat <- lme4::sleepstudy
fit <- lme4::lmer(Reaction ~ Days + (1|Subject), data = lme4::sleepstudy)
add_probs(dat, fit, q = 300)
add_probs(dat, fit, q = 300, type = "parametric", includeRanef = FALSE, comparison = ">")
add_probs(dat, fit, q = 300, type = "sim")

}
\seealso{
\code{{\link{add_ci.lmerMod}}} for confidence intervals
    for \code{lmerMod} objects. \code{\link{add_pi.lmerMod}} for
    prediction intervals of \code{lmerMod} objects, and
    \code{\link{add_quantile.lmerMod}} for response quantiles of
    \code{lmerMod} objects.
}
