% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_probs_survreg.R
\name{add_probs.survreg}
\alias{add_probs.survreg}
\title{Confidence Intervals for the Survivor Function of Parametric Survival Models}
\usage{
\method{add_probs}{survreg}(tb, fit, q, name = NULL,
  yhatName = "median_pred", comparison = "<", method = "parametric",
  confint = TRUE, alpha = 0.05, nSims = 2000, ...)
}
\arguments{
\item{tb}{A tibble or data frame of new data.}

\item{fit}{An object of class \code{survreg}. Predictions are made
with this object.}

\item{q}{A double. A quantile of the survival time distribution. In
survival applications this is the time of event.}

\item{name}{\code{NULL} or a string. If \code{NULL}, probabilities
automatically will be named by \code{add_probs()}, otherwise,
the probabilities will be named \code{name} in the returned
tibble}

\item{yhatName}{A string. Name of the vector of predictions.}

\item{comparison}{A character vector of length one. If
\code{comparison = "<"}, then \eqn{Pr(Y|X < q)} is
calculated. If \code{comparison = ">"}, the survivor function
at time \code{q} is calculated.}

\item{method}{A string. One of either \code{"parametric"} or
\code{"bootstrap"}.}

\item{confint}{A logical. If \code{TRUE}, confidence intervals for
the estimated probabilities will be calculated and appended to
\code{tb}.}

\item{alpha}{A number. Control the confidence level of the
confidence intervals if \code{confint = TRUE}.}

\item{nSims}{Number of simulations used if \code{method = "boot"}}

\item{...}{Additional arguments.}
}
\value{
A tibble, \code{tb}, with predicted medians, probabilities,
    and confidence intervals for predicted probabilities attached.
}
\description{
This function is one of the methods of \code{add_probs} and is
automatically called when an object of class \code{survreg} is
passed to \code{add_probs}.
}
\details{
Confidence intervals may be produced for estimated probabilities of
accelerated failure time models. Presently, confidence intervals
may be computed for lognormal, weibull, exponential, and
loglogistic failure time models. If \code{comparison = "<"},
confidence intevals are made for the probability that a failure
will be observed before \code{q}. Similarly, if \code{comparison =
">"}, confidence intervals will be formed for the probability that
a unit fails after \code{q}. In the survival literature,
\code{compaison = ">"} corresponds to estimating the survivor
function, \emph{S(q)}.

Confidence intervals may be produced paramerically via the Delta
Method, or computationally through a nonparametric (case
resampling) bootstrap procedure. Simulations show that under a mild
to moderate amount of censoring, these methods perform
similarly. Thus, the parametric method may be preferred in many
standard applications.

In both methods, the logistic transformation is applied to ensure
that confidence interval bounds lie between \eqn{0} and \eqn{1}.

Note: Due to a limitation, the \code{Surv} object must be specified in
\code{survreg} function call. See the examples section for one way
to do this.

Note: \code{add_probs.survreg} cannot inspect the convergence of
\code{fit}. Poor maximum likelihood estimates will result in poor
confidence intervals. Inspect any warning messages given from
\code{survreg}.
}
\examples{
## Define a data set.
tb <- survival::stanford2
## remove a covariate with missing values.
tb <- tb[, 1:4]
## next, create the Surv object inside the survreg call:
fit <- survival::survreg(survival::Surv(time, status) ~ age + I(age^2),
                         data = tb, dist = "lognormal")
## Calculate the level 0.75 quantile wit CIs for that quantile
add_probs(tb, fit, q = 500, name = c("Fhat", "lwr", "upr"))

## Try a weibull model for the same data:
fit2 <- survival::survreg(survival::Surv(time, status) ~ age + I(age^2),
                          data = tb, dist = "weibull")
## Calculate the level 0.75 quantile with CIs for the quantile
add_probs(tb, fit2, q = 500, name = c("Fhat", "lwr", "upr"))


}
\references{
For the logistic transformation of estimated probabilities and error bounds:
Meeker, William Q., and Luis A. Escobar. Statistical methods for reliability data. John Wiley & Sons, 2014. (Chapter 8)

For a discussion of forming confidence intervals for survival probabilities:
Harrell, Frank E. Regression modeling strategies. Springer, 2015. (Chapter 17)
}
\seealso{
\code{\link{add_ci.survreg}} for confidence intervals for
    \code{survreg} objects, \code{\link{add_pi.survreg}} for
    prediction intervals of \code{survreg} objects, and
    \code{\link{add_quantile.survreg}} for response quantiles of
    \code{survreg} objects.
}
