% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_quantile_lm.R
\name{add_quantile.lm}
\alias{add_quantile.lm}
\title{Quantiles for the Response of a Linear Model}
\usage{
\method{add_quantile}{lm}(tb, fit, p, name = NULL, log_response = FALSE,
  ...)
}
\arguments{
\item{tb}{A tibble or Data Frame.}

\item{fit}{An object of class lm. Predictions are made with this
object.}

\item{p}{A real number between 0 and 1. Sets the probability for
which to calculate the quantiles.}

\item{name}{NULL or character vector of length one. If \code{NULL},
quantiles will automatically be named by \code{add_quantile},
otherwise, they will be named \code{name}}

\item{log_response}{logical. If TRUE, quantiles will be generated
for the prediction made with a log-linear model: \eqn{\log(Y) =
X\beta + \epsilon}. These quantiles will be on the scale of the
original response, Y.}

\item{...}{Additional arguments.}
}
\value{
A tibble, \code{tb}, with predicted values, upper and lower
    prediction bounds attached.
}
\description{
This function is one of the methods for \code{add_quantile}. It is
automatically called when \code{add_quantile} is applied to objects
of class \code{lm}.
}
\details{
Quantiles for linear models are determined parametrically, more or
less the same way that prediction intervals for this class of
models is determined.
}
\examples{
fit <- lm(dist ~ speed, data = cars)
add_quantile(cars, fit, p = 0.7)
add_quantile(cars, fit, p = 0.7, name = "my_quantile")

}
\seealso{
\code{{\link{add_ci.lm}}} for confidence intervals for
    \code{lm} objects. \code{\link{add_pi.lm}} for prediction
    intervals of \code{lm} objects, and \code{\link{add_probs.lm}}
    for response probabilities of \code{lm} objects.
}
