% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_quantile_lmer.R
\name{add_quantile.lmerMod}
\alias{add_quantile.lmerMod}
\title{Quantiles for the Response of a Linear Mixed Model}
\usage{
\method{add_quantile}{lmerMod}(tb, fit, p, name = NULL, includeRanef = TRUE,
  type = "boot", nSims = 200, log_response = FALSE, yhatName = "pred",
  ...)
}
\arguments{
\item{tb}{A tibble or data frame of new data.}

\item{fit}{An object of class \code{lm}. Predictions are made with
this object.}

\item{p}{A real number between 0 and 1.  Determines the probability
level of the quantiles.}

\item{name}{\code{NULL} or a string. If \code{NULL}, quantiles will
automatically be named by \code{add_quantile}, otherwise, they
will be named \code{name}}

\item{includeRanef}{Should the quantiles be calculated condition on
the random effects? If \code{TRUE}, quantiles will be
calculated at the "group level". Otherwise, quantiles will be
calculated at the "population level".}

\item{type}{A string. Options are \code{"parametric"},
\code{"boot"}, or \code{"sim"}.}

\item{nSims}{A positive integer. Set the number of simulations to
perform. Only applied when \code{type = "boot"} or \code{type =
"sim"}}

\item{log_response}{A logical. Set to \code{TRUE} if the model is a
log-linear mixed model: \eqn{\log(Y) = X\beta + Z\gamma + \epsilon}.}

\item{yhatName}{A string. Determines the name of the column of
predictions.}

\item{...}{Additional arguments.}
}
\value{
A tibble, \code{tb}, with predicted values and level-p
    quantiles attached.
}
\description{
This function is one of the methods for \code{add_quantile} and is
automatically called when \code{add_quantile} is applied to an
object of class \code{lmerMod}.
}
\details{
\code{add_qauntile.lmerMod} may use three different methods for
determining quantiles: a parametric method, a simulation method
(via \code{merTools::predictInterval}), or a parametric bootstrap
method (via \code{lme4::simulate}). The default and recommended
method is parametric bootstrap, which corresponds to setting
\code{type = "boot"}. Only use the parametric method (\code{type =
"parametric"}) if \code{fit} is a random intercept model
e.g. \code{fit = lmer(y ~ x + (1|group))}.
}
\examples{
dat <- lme4::sleepstudy

# Fit a random intercept model
fit <- lme4::lmer(Reaction ~ Days + (1|Subject), data = lme4::sleepstudy)

# Using the parametric bootstrap: Given the model fit, what value
# of reaction time do we expect half of new reaction times to fall
# under?
add_quantile(dat, fit, p = 0.5)

# Using the parametric method:
# As above, but we ignore the random effects
add_quantile(dat, fit, p = 0.5, type = "parametric", includeRanef = FALSE)

# Using the simulation method, and giving the vector of quantiles a
# custom name:
add_quantile(dat, fit, p = 0.5, type = "sim", name = "my_quantile", nSims = 1000)

}
\seealso{
\code{\link{add_ci.lmerMod}} for confidence intervals
    for \code{lmerMod} objects. \code{\link{add_pi.lmerMod}} for
    prediction intervals of \code{lmerMod} objects, and
    \code{\link{add_probs.lmerMod}} for response probabilities of
    \code{lmerMod} objects.
}
