% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_quantile_survreg.R
\name{add_quantile.survreg}
\alias{add_quantile.survreg}
\title{Confidence Intervals for Survival Time Quantiles of a Parametric
Survival Model}
\usage{
\method{add_quantile}{survreg}(tb, fit, p = 0.5, name = NULL,
  yhatName = "median_pred", confint = TRUE, alpha = 0.1,
  method = "parametric", nSims = 2000, ...)
}
\arguments{
\item{tb}{A tibble or data frame of new data.}

\item{fit}{An object of class \code{survreg}. Predictions are made
with this object.}

\item{p}{A real number between 0 and 1. Sets the probability level
of the quantiles.}

\item{name}{\code{NULL} or a string. If \code{NULL}, quantiles
automatically will be named by \code{add_quantile}, otherwise,
they will be named \code{name}.}

\item{yhatName}{A string. Name of the vector of predictions.}

\item{confint}{A logical. If \code{TRUE}, confidence intervals for
the quantiles are also appended to \code{tb}.}

\item{alpha}{A number. Controls the confidence level of the
confidence intervals if \code{confint = TRUE}.}

\item{method}{A string. One of either \code{"parametric"} or
\code{"boot"}.}

\item{nSims}{A positive integer. Set the number of simulated draws
to use.}

\item{...}{Additional arguments.}
}
\value{
A tibble, \code{tb}, with predicted medians, level \emph{p}
    quantiles attached and confidence intervals attached.
}
\description{
This function is one of the methods of \code{add_quantile} and is
automatically called when an object of class \code{survreg} is
passed to \code{add_quantile}.
}
\details{
Details here.
}
\examples{

}
\references{

}
\seealso{
\code{\link{add_ci.survreg}} for confidence intervals for
    \code{survreg} objects, \code{\link{add_pi.survreg}} for
    prediction intervals of \code{survreg} objects, and
    \code{\link{add_probs.survreg}} for response probabilities of
    \code{survreg} objects.
}
