\name{CItools-package}
\alias{CItools-package}
\alias{CItools}
\docType{package}
\title{
Computation, comparison and sensitivity analysis of composite indicators of performance
}
\description{
This package provides functions to compute, compare and test the sensitivity of index based on composite indicators of performance.
}
\details{
\tabular{ll}{
Package: \tab CItools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-06-21\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
A composite indicator is formed when individual indicators are compiled into a single index on the basis of an underlying model.
Using this package, user can upload his own dataset of composite indicators and then compute, compare and test the sensitivity of index.
While computing the index, the user can give options to perform imputation of missing data, cleaning, normalization, weighting and aggregation.
}
\author{
David Ardia, Kris Boudt, Valentin Todorov and Keshav Pratap

Maintainer: Kris Boudt <Kris.Boudt@econ.kuleuven.be>
}
\references{
OECD (2008) Handbook on Constructing Composite Indicators, Paris: OECD.

Saisana, M., Saltelli, A. and Tarantola, S. (2005) 'Uncertainty and sensitivity analysis techniques as tools for the quality assessment of composite indicators.' Journal of Royal Statistical Society A 168, 307-323. 
}

\keyword{ package }

