% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{compute_density}
\alias{compute_density}
\title{Calcula densidad de cotización sobre un periodo base}
\usage{
compute_density(data, start, end)
}
\arguments{
\item{data}{Cuenta individual con variables \code{CVE_NSS}, \code{FEC_MOV_INI}, \code{FEC_MOV_FIN}}

\item{start}{Inicio del periodo base}

\item{end}{Fin del periodo base}
}
\value{
Una \code{\link[tibble:tibble]{tibble::tibble()}} con una variable adicional que calcula la densidad
de cotización aportada por cada movimiento afiliatorio.
}
\description{
Calcula densidad de cotización sobre un periodo base
}
