% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{is_work_mode}
\alias{is_work_mode}
\title{¿Modalidad considerada trabajo?}
\usage{
is_work_mode(mode)
}
\arguments{
\item{mode}{Un vector de modalidades}
}
\value{
Resultado lógico
}
\description{
¿Modalidad considerada trabajo?
}
\examples{
a <- c("10", "31")
is_work_mode(a)
}
