% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_intervals}
\alias{plot_intervals}
\title{Graficar la vida laboral de un NSS}
\usage{
plot_intervals(data, NSS)
}
\arguments{
\item{data}{Cuenta individual con variables \code{CVE_NSS}, \code{FEC_MOV_INI},
\code{FEC_MOV_FIN}, \code{CVE_PATRON}, \code{SAL_BAS}}

\item{NSS}{caracter. Número de seguridad social}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} class element.
}
\description{
Graficar la vida laboral de un NSS
}
