\name{tai}
\alias{tai}
\docType{data}
\title{
TAI Data
}
\description{
List of individual indicators of the Technology Achievement Index
}
\usage{data(tai)}
\format{
  A data frame with 72 observations on the following 9 variables.
  \describe{
    \item{\code{country}}{Names of the countries for which TAI index is calculated}
    \item{\code{patents}}{Patents granted per 1000000 people: Number of patents granted to residents, to reflect the current level of
invention activities (1998)}
    \item{\code{royalties}}{US $ per 1000 people: Receipts of royalty and license fees from abroad per capita, so as to reflect the stock of successful innovations of the past that are still useful and hence have market value (1999)}
    \item{\code{internet}}{Internet hosts per 1000 people: Diffusion of the Internet, which is indispensable to participation in the network age (2000)}
    \item{\code{exports}}{Percent: Exports of high and medium technology products as a share of total goods exports (1999)}
    \item{\code{telephones}}{Telephone lines per 1000 people (log): Number of telephone lines (mainline and cellular), which represents old innovation needed to use newer technologies and is also pervasive input to a multitude of human activities (1999)}
    \item{\code{electricity}}{kWh per capita (log) Electricity consumption, which represents old innovation needed to use newer technologies and is also pervasive input to a multitude of human activities (1998)}
    \item{\code{schooling}}{Years: Mean years of schooling (age 15 and above), which represents the basic education needed to develop cognitive skills (2000)}
    \item{\code{university}}{percent: Gross enrolment ratio of tertiary students enrolled in science, mathematics and engineering, which reflects the human skills needed to create and absorb innovations (1995-1997)}
  }
}

\source{
OECD (2008) Handbook on Constructing Composite Indicators, Paris: OECD. pg 152-153
}

\keyword{datasets}
