% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{tidy_intervals}
\alias{tidy_intervals}
\title{Recorta eventos de la Cuenta Individual para calcular la densidad de
cotización}
\usage{
tidy_intervals(start, end)
}
\arguments{
\item{start}{Fecha de movimiento de inicio}

\item{end}{Fecha de movimiento de fin}
}
\value{
Un intervalo \code{\link[lubridate:interval]{lubridate::interval()}}
}
\description{
Recorta eventos de la Cuenta Individual para calcular la densidad de
cotización
}
