% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{working_days}
\alias{working_days}
\title{Calcula los días de trabajo para un movimiento afiliatorio.}
\usage{
working_days(interval, start = NULL, end = NULL)
}
\arguments{
\item{interval}{un vector de longitud \code{n} de objetos clase \code{\link[lubridate:interval]{lubridate::interval()}}}

\item{start}{Alternativamente, la fecha de inicio del movimiento afiliatorio.}

\item{end}{Fecha de fin del movimiento afiliatorio.}
}
\value{
Un vector longitud \code{n} de clase \code{numeric} que corresponde a los días
trabajados en cada movimiento afiliatorio.
}
\description{
Calcula los días de trabajo para un movimiento afiliatorio.
}
