% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ci_survreg.R
\name{add_ci.survreg}
\alias{add_ci.survreg}
\title{Confidence Intervals for the Expected Response of Parametric
Survival Models.}
\usage{
\method{add_ci}{survreg}(tb, fit, alpha = 0.1, names = NULL,
  yhatName = "mean_pred", method = "parametric", nSims = 2000, ...)
}
\arguments{
\item{tb}{A tibble or data frame of new data.}

\item{fit}{An object of class \code{survreg}. Predictions are made
with this object.}

\item{alpha}{A number between 0 and 1. 1 - \code{alpha} is the
confidence level of the intervals.}

\item{names}{\code{NULL} or a string. If \code{NULL}, quantiles
automatically will be named by \code{add_quantile}, otherwise,
they will be named \code{name}.}

\item{yhatName}{A string. Name of the vector of predictions.}

\item{method}{A string. One of either \code{"parametric"} or
\code{"boot"}.}

\item{nSims}{A positive integer. Set the number of simulated draws
to use.}

\item{...}{Additional arguments.}
}
\value{
A tibble, \code{tb}, with predicted expected values and level
    \emph{alpha} confidence levels attached.
}
\description{
This function is one of the methods of \code{add_ci}.
}
\examples{

}
\references{

}
\seealso{
\code{\link{add_quantile.survreg}} for quantiles of
    \code{survreg} objects, \code{\link{add_pi.survreg}} for
    prediction intervals of \code{survreg} objects, and
    \code{\link{add_probs.survreg}} for response probabilities of
    \code{survreg} objects.
}
