% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pi_glm.R
\name{add_pi.glm}
\alias{add_pi.glm}
\title{Prediction Intervals for Generalized Linear Models}
\usage{
\method{add_pi}{glm}(tb, fit, alpha = 0.05, names = NULL, type = "sim",
  nSims = 1000, ...)
}
\arguments{
\item{tb}{A tibble or Data Frame.}

\item{fit}{An object of class \code{glm}.}

\item{alpha}{A real number between 0 and 1. Controls the confidence
level of the interval estimates.}

\item{names}{NULL or character vector of length two. If
\code{NULL}, prediction bounds will automatically be named by
\code{add_pi}, otherwise, the lower prediction bound will be
named \code{names[1]} and the upper prediction bound will be
named \code{names[2]}.}

\item{type}{A string. Currently must be \code{"sim"}.}

\item{nSims}{A positive integer. Determines the number of
simulations to run.}

\item{...}{Additional arguments.}
}
\value{
A tibble, \code{tb}, with predicted values, upper and lower
    prediction bounds attached.
}
\description{
This function is one of the methods for \code{add_pi}, and is
called automatically when \code{add_pi} is used on a \code{fit} of
class \code{glm}.
}
\details{
Prediction Intervals are generated through simulation via the
function \code{sim} in the package \code{arm}. At the moment, only
prediction intervals for Poisson GLMs with the log link function
are supported. Note that if the response is count data, e.g. the
response is modeled with a Poisson regression, prediction intervals
are only approximate.
}
\examples{
# Poisson Regression
fit <- glm(dist ~ speed, data = cars, family = "poisson")
add_pi(cars, fit)
add_pi(cars, fit, alpha = 0.5)
add_pi(cars, fit, alpha = 0.5, names = c("lwr", "upr"))

}
\seealso{
\code{{\link{add_ci.glm}}} for confidence intervals for
    \code{glm} objects. \code{\link{add_probs.glm}} for conditional
    probabilities of \code{glm} objects, and
    \code{\link{add_quantile.glm}} for response quantiles of
    \code{glm} objects.
}
