% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pi_lmer.R
\name{add_pi.lmerMod}
\alias{add_pi.lmerMod}
\title{Prediction Intervals for the Response of Linear Mixed Models}
\usage{
\method{add_pi}{lmerMod}(tb, fit, alpha = 0.05, type = "parametric",
  includeRanef = TRUE, names = NULL, log_response = FALSE, nSims = 200,
  yhatName = "pred")
}
\arguments{
\item{tb}{A tibble or Data Frame.}

\item{fit}{An object of class \code{lmerMod}.}

\item{alpha}{A real number between 0 and 1. Controls the confidence
level of the interval estimates.}

\item{type}{A string, either \code{"parametric"}, \code{"sim"},
\code{"sim_lme4"}.}

\item{includeRanef}{A logical. Set whether the predictions and
intervals should be made conditional on the random effects. If
\code{FALSE}, random effects will not be included.}

\item{names}{NULL or character vector of length two. If
\code{NULL}, prediction bounds will automatically be named by
\code{add_pi}, otherwise, the lower prediction bound will be
named \code{names[1]} and the upper prediction bound will be
named \code{names[2]}.}

\item{log_response}{A logical, indicating if the response is on log
scale in the model. If \code{TRUE}, prediction intervals will
be returned on the response scale.}

\item{nSims}{A positive integer. If \code{type = "sim"} or
\code{type = "sim_lme4"}, \code{nSims} will determine the
number of simulated draws to make.}

\item{yhatName}{A character vector of length one. Names of the
column of predictions from \code{fit} on \code{tb}.}
}
\value{
A tibble, \code{tb}, with predicted values, upper and lower
    prediction bounds attached.
}
\description{
This function is one of the methods for \code{add_pi}, and is
called automatically when \code{add_pi} is used on a \code{fit} of
class \code{lmerMod}. It is recommended that one use parametric
prediction intervals when modeling with a random intercept
LMM. Otherwise prediction intervals may be simulated using one of
two methods. \code{"sim"} indicates that \code{predictInterval}
from \code{merTools} should be used to simulate responses to form
prediction intervals, and \code{"sim_lme4"} indicates that
\code{simulate.merMod} should be used to simulate predictions.
}
\examples{
\dontrun{
data(sleepstudy) ## included in lme4
tb <- sleepstudy
fit1 <- lmer(Reaction ~ Days + (1 | Subject), data = sleepstudy)
add_pi.lmerMod(tb, fit1)
add_pi.lmerMod(tb, fit1, type = "sim")
add_pi.lmerMod(tb, fit1, type = "sim_lme4")
}

}
