% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pi_survreg.R
\name{add_pi.survreg}
\alias{add_pi.survreg}
\title{Prediction Intervals for Survival Models}
\usage{
\method{add_pi}{survreg}(tb, fit, alpha = 0.05, names = NULL,
  yhatName = "pred", nSims = 2000, bootData = FALSE, ...)
}
\arguments{
\item{tb}{A tibble or data frame of new data.}

\item{fit}{An object of class \code{survreg}.}

\item{alpha}{A real number between 0 and 1. Controls the confidence
level of the interval estimates.}

\item{names}{\code{NULL} or character vector of length two. If
\code{NULL}, prediction bounds automatically will be named by
\code{add_pi}, otherwise, the lower prediction bound will be
named \code{names[1]} and the upper prediction bound will be
named \code{names[2]}.}

\item{yhatName}{A string. Name of the predictions vector.}

\item{nSims}{A positive integer. Determines the number of}

\item{bootData}{A logical. If \code{TRUE} a matrix is bootstrap
replicates is returned.  simulations to run.}

\item{...}{Additional arguments.}
}
\value{
A tibble, \code{tb}, with predicted values, upper and lower
    prediction bounds attached.
}
\description{
This function is one of the methods for \code{add_pi}, and is
called automatically when \code{add_pi} is used on a \code{fit} of
class \code{survreg}.
}
\examples{

}
\seealso{
\code{\link{add_ci.survreg}} for confidence intervals for
    \code{survreg} objects, \code{\link{add_probs.survreg}} for
    conditional probabilities of \code{survreg} objects, and
    \code{\link{add_quantile.survreg}} for response quantiles of
    \code{survreg} objects.
}
