% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_probs_glm.R
\name{add_probs.glm}
\alias{add_probs.glm}
\title{Response Probabilities for Generalized Linear Models}
\usage{
\method{add_probs}{glm}(tb, fit, q, name = NULL, comparison = "<",
  nSims = 200, ...)
}
\arguments{
\item{tb}{A tibble or Data Frame on which to append probabilities}

\item{fit}{An object of class \code{glm}. Predictions are made with
this object.}

\item{q}{A double. A quantile of the response distribution.}

\item{name}{NULL or character vector of length one. If \code{NULL},
probabilities will automatically be named by
\code{add_probs()}, otherwise, the probabilities will be named
\code{name} in the returned tibble}

\item{comparison}{A character vector of length one. If
\code{comparison = "<"}, then Pr(Y|X < q) is calculated. Any
comparison is allowed in poisson regression, but only certain
comparisons may be made in Logistic regression. See the Details
section.}

\item{nSims}{A positive integer.}

\item{...}{Additional arguments.}
}
\value{
A tibble, \code{tb}, with predicted values and
    probabilities attached.
}
\description{
This is the method \code{add_probs} uses if the model fit is an
object of class \code{glm}. Probabilities are determined through
simulation, using the same method as \code{add_pi.glm}. Currently,
only logistic and poisson regression are supported.
}
\details{
Any of the five comparisons, \code{comparison = "<"}, \code{">"},
\code{"="}, \code{"<="}, or \code{">="} may be made for a Poisson
Model. For logistic regression, the comparison statement must be
equivalent to \eqn{Pr(Y|x = 0)} or \eqn{Pr(Y|x = 1)}. If
\code{add_probs} is called on a Poisson model, a simulation is
preformed using the function \code{sim} from the package \code{arm}.
}
\examples{
fit <- glm(dist ~ speed, data = cars, family = "poisson")
add_probs(cars, fit, q = 20)
add_probs(cars, fit, q = 30, comparison = ">")
add_probs(cars, fit, q = 30, comparison = ">=")

}
\seealso{
\code{{\link{add_ci.glm}}} for confidence intervals for
    \code{glm} objects. \code{\link{add_pi.glm}} for prediction
    intervals of \code{glm} objects, and
    \code{\link{add_quantile.glm}} for response quantiles of
    \code{glm} objects.
}
