% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_probs_lm.R
\name{add_probs.lm}
\alias{add_probs.lm}
\title{Event Probabilities for Linear Models}
\usage{
\method{add_probs}{lm}(tb, fit, q, name = NULL, comparison = "<",
  log_response = FALSE)
}
\arguments{
\item{tb}{A tibble or Data Frame on which to append probabilities}

\item{fit}{An object of class lm. Predictions are made with this
object.}

\item{q}{A double. A quantile of the response distribution.}

\item{name}{NULL or character vector of length one. If \code{NULL},
probabilities will automatically be named by
\code{add_probs()}, otherwise, the probabilities will be named
\code{name} in the returned tibble}

\item{comparison}{A character vector of length one. If
\code{comparison = "<"}, then Pr(Y|x < q) is calculated for
each observation in \code{tb}. Must be "<" or ">" for linear,
log-linear and linear mixed models. If \code{fit} is a glm,
then \code{comparison} may also be "<=", ">=", or "=".}

\item{log_response}{A logical. If TRUE, then the quantile
\code{t} will be transformed to the log scale.}
}
\value{
A tibble, \code{tb}, with predicted values and
    probabilities attached.
}
\description{
This is the method \code{add_probs} uses if the model fit is
linear.  Probabilities are calculated using the t-distribution.
}
\examples{
# linear regression
fit1 <- lm(dist ~ speed, data = cars)
# Calculate Pr(dist < 40) for each observation in cars
add_probs(cars, fit1, q = 40)
add_probs(cars, fit1, q = 50, comparison = ">")

}
