% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_probs_lmer.R
\name{add_probs.lmerMod}
\alias{add_probs.lmerMod}
\title{Response Probabilities for Linear Mixed Models}
\usage{
\method{add_probs}{lmerMod}(tb, fit, q, type = "parametric",
  includeRanef = TRUE, name = NULL, comparison = "<", nSims = 200,
  log_response = FALSE)
}
\arguments{
\item{tb}{A tibble or Data Frame.}

\item{fit}{An object of class \code{lmerMod}.}

\item{q}{A double. A quantile of the response variable}

\item{type}{A string, either \code{"parametric"} , \code{"sim"}, or
\code{"sim_lme4"}.}

\item{includeRanef}{A logical. Set whether the predictions and
intervals should be made conditional on the random effects. If
\code{FALSE}, random effects will not be included.}

\item{name}{NULL or character vector of length two. If \code{NULL},
probabilities will automatically be named by \code{add_pi},
otherwise, the probabilities will be named \code{name} in the
returned tibble.}

\item{comparison}{A character vector of length one. If
\code{comparison = "<"}, then Pr(Y|x < q) is calculated for
each observation in \code{tb}. Must be "<" or ">" for linear,
log-linear and linear mixed models. If \code{fit} is a glm,
then \code{comparison} may also be "<=", ">=", or "=".}

\item{nSims}{A positive integer. If \code{type = "sim"}
\code{nSims} will determine the number of simulated draws to
make.}

\item{log_response}{A logical. Set to \code{TRUE} if your model is
a log-linear mixed model.}
}
\value{
A tibble, \code{tb}, with predictions and probabilities
    attached.
}
\description{
This function is one of the methods for \code{add_probs}, and is
called automatically when \code{add_probs} is used on a \code{fit}
of class \code{lmerMod}. It is recommended that one calculate
parametric probabilties when modeling with a random intercept
mixed model. Otherwise, probabilities may be simulated.
}
