% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_probs.R
\name{add_probs}
\alias{add_probs}
\title{Add Probabilities to Data Frames}
\usage{
add_probs(tb, fit, q, name = NULL, comparison, ...)
}
\arguments{
\item{tb}{A tibble or Data Frame on which to append probabilities}

\item{fit}{An object of class lm, glm, or lmerMod. Predictions are
made with this object.}

\item{q}{A double. A quantile of the response distribution.}

\item{name}{NULL or character vector of length one. If \code{NULL},
probabilities will automatically be named by
\code{add_probs()}, otherwise, the probabilities will be named
\code{name} in the returned tibble}

\item{comparison}{A character vector of length one. If
\code{comparison = "<"}, then Pr(Y|x < q) is calculated for
each observation in \code{tb}. Default is "<". Must be "<" or
">" for linear, log-linear and linear mixed models. If
\code{fit} is a glm, then comparison may also be "<=", ">=", or
"=".}

\item{...}{Additional arguments}
}
\value{
A tibble, \code{tb}, with predicted values and
    probabilities attached.
}
\description{
This is a generic function to append event probabilities to a data
frame. An event probability, such as Pr(Response|Covariates < 10),
is generated for the fitted value of each observation in
\code{tb}. These probabilities are then appended to \code{tb} and
returned to the user as a tibble.
}
