% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_probs.R
\name{add_probs}
\alias{add_probs}
\title{Add Regression Probabilities to Data Frames}
\usage{
add_probs(tb, fit, q, name = NULL, comparison, ...)
}
\arguments{
\item{tb}{A tibble or Data Frame on which to append probabilities}

\item{fit}{An object of class lm, glm, or lmerMod. Predictions are
made with this object.}

\item{q}{A double. A quantile of the response distribution.}

\item{name}{NULL or character vector of length one. If \code{NULL},
probabilities will automatically be named by
\code{add_probs()}, otherwise, the probabilities will be named
\code{name} in the returned tibble}

\item{comparison}{A character vector of length one. If
\code{comparison = "<"}, then Pr(Y|x < q) is calculated for
each observation in \code{tb}. Default is "<". Must be "<" or
">" for linear, log-linear and linear mixed models. If
\code{fit} is a glm, then comparison may also be "<=", ">=", or
"=".}

\item{...}{Additional arguments}
}
\value{
A tibble, \code{tb}, with predicted values and
    probabilities attached.
}
\description{
This is a generic function to append response level probabilities
to a data frame. A response level probability, such as
Pr(Response|Covariates < 10), is generated for the fitted value of
each observation in \code{tb}. These probabilities are then
appended to \code{tb} and returned to the user as a tibble.
}
\details{
For more specific information about the arguments that are useful
in each method, consult

\itemize{
  \item \code{\link{add_probs.lm}} for linear regression response probabilities
  \item \code{\link{add_probs.glm}} for generalized linear regression response probabilities
  \item \code{\link{add_probs.lmerMod}} for linear mixed models response probabilities
}

The probabilities that \code{add_probs} calculates are on the
distribution of \eqn{Y|x}, not \eqn{E[Y|x]}. That is, they use the
same distribution as a prediction interval, not the distribution of
a confidence interval.
}
\examples{
fit <- lm(dist ~ speed, data = cars)
add_probs(cars, fit, q = 20)
fit2 <- glm(dist ~ speed, family = "poisson", data = cars) 
add_probs(cars, fit2, q = 20)
fit3 <- lme4::lmer(Reaction ~ Days + (1|Subject), data = lme4::sleepstudy)
add_probs(lme4::sleepstudy, fit3, q = 300)


}
\seealso{
\code{\link{add_ci}} for confidence intervals,
    \code{\link{add_quantile}} for response level quantiles, and
    \code{\link{add_pi}} for prediction intervals
}
