% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_probs_survreg.R
\name{add_probs.survreg}
\alias{add_probs.survreg}
\title{Confidence Intervals for the Survivor Function of Parametric Survival Models}
\usage{
\method{add_probs}{survreg}(tb, fit, q, confint = TRUE, alpha = 0.05,
  name = NULL, yhatName = "pred", method = "parametric",
  comparison = "<", nSims = 2000, ...)
}
\arguments{
\item{tb}{A tibble or data frame of new data.}

\item{fit}{An object of class \code{survreg}. Predictions are made
with this object.}

\item{q}{A double. A quantile of the response distribution. In
survival applications this is the time of event.}

\item{confint}{A logical. If \code{TRUE}, confidence intervals for
the estimated probabilities will be calculated and appended to
\code{tb}.}

\item{name}{\code{NULL} or a string. If \code{NULL}, probabilities
automatically will be named by \code{add_probs()}, otherwise,
the probabilities will be named \code{name} in the returned
tibble}

\item{yhatName}{A string. Name of the vector of predictions.}

\item{method}{A string. One of either \code{"parametric"} or
\code{"bootstrap"}.}

\item{comparison}{A character vector of length one. If
\code{comparison = "<"}, then \eqn{Pr(Y|X < q)} is
calculated. If \code{comparison = ">"}, the survivor function
at time \code{q} is calculated.}

\item{nSims}{Number of simulations used for bootstrap confidence
intervals.}

\item{...}{Additional arguments.}
}
\value{
A tibble, \code{tb}, with predicted values and
    probabilities attached.
}
\description{
TODO add a description
}
\details{
TODO add details
}
\examples{

}
\seealso{
\code{\link{add_ci.survreg}} for confidence intervals for
    \code{survreg} objects, \code{\link{add_pi.survreg}} for
    prediction intervals of \code{survreg} objects, and
    \code{\link{add_quantile.survreg}} for response quantiles of
    \code{survreg} objects.
}
