% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_quantile_glm.R
\name{add_quantile.glm}
\alias{add_quantile.glm}
\title{Quantiles for the Response of a Generalized Linear Model}
\usage{
\method{add_quantile}{glm}(tb, fit, p, name = NULL, nSims = 200, ...)
}
\arguments{
\item{tb}{A tibble or Data Frame.}

\item{fit}{An object of class lm. Predictions are made with this
object.}

\item{p}{A real number between 0 and 1. Sets the probability for
which to calculate the quantiles.}

\item{name}{NULL or character vector of length one. If \code{NULL},
quantiles will automatically be named by \code{add_quantile},
otherwise, they will be named \code{name}}

\item{nSims}{A positive integer. Set the number of simulated draws
to use.}

\item{...}{Additional arguments.}
}
\value{
A tibble, \code{tb}, with predicted values and level
    \emph{p} quantiles attached.
}
\description{
This function is one of the methods for
\code{add_quantile}. Currently, you can only use this function to
compute the quantiles of the response of a Poisson regression with
the log link function.
}
\details{
Quantiles are GLMs are determined by \code{add_quantile} through a
simulation using the function \code{sim} from the package
\code{arm}.
}
\examples{
fit <- glm(dist ~ speed, data = cars, family = "poisson")
add_quantile(cars, fit, p = 0.3)
add_quantile(cars, fit, p = 0.5, name = "my_quantile", nSims = 300)

}
\seealso{
\code{{\link{add_ci.lm}}} for confidence intervals for
    \code{lm} objects. \code{\link{add_pi.lm}} for prediction
    intervals of \code{lm} objects, and \code{\link{add_probs.lm}}
    for response probabilities of \code{lm} objects.
}
