% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citools-package.R
\docType{package}
\name{citools-package}
\alias{citools}
\alias{citools-package}
\title{citools: Tools for Cuenta Individual}
\description{
Work with data from Cuenta Individual.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/estebandegetau/citools}
  \item Report bugs at \url{https://github.com/estebandegetau/citools/issues}
}

}
\author{
\strong{Maintainer}: Esteban Degetau \email{estebandegetau@gmail.com} (\href{https://orcid.org/0009-0004-4095-8819}{ORCID})

}
\keyword{internal}
